Refactor
Rename FileDemo to Museum

Take all the GUI set up code and move it into a new file called UI.java

UI.java defines a new class called UI which represents the user
interface of our museum. The main Museum class now instantiates a copy
of this UI and the UI() constructor sets up all the buttons, text
fields, etc how we need them.

The UI() constructor also accepts an ActionListener and
ListSelectionListener and stores them in a private variable, all
buttons/filters, etc that need to emit actions have these added as
action listeners or list selection listeners. The UI elements that need
to emit action events are made public so that the Museum class (or
whatever actionlistener is passed in) can specify them in the action
handling code.

When Museum calls the UI constructor, it passes itself in as the
action/list selection listener. The code for the actual action handling
has not been added yet.